Update README.md
Thanks for this excellent series of Best Practices. This is an important document that is a living and dynamic resource. I would love to assist in maintaining and updating this information to also include additional references and information. I have also edited the copy for grammer and spelling, and made several corrections to the narrative for the sake of efficiency and accuracy.